change AttendanceTime type.

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.app/src/main/java/nts/uk/ctx/at/shared/app/employment/statutory/worktime/shared/FlexMonth.java
@@ -11,13 +11,13 @@ import lombok.Data;
  */
 @Data
 public class FlexMonth {
-	
+
 	/** The month. */
 	private int month;
-	
+
 	/** The statutory time. */
-	private int statutoryTime;
-    
-    /** The specified time. */
-    private int specifiedTime;
+	private Long statutoryTime;
+
+	/** The specified time. */
+	private Long specifiedTime;
 }
